Improve deleteBlogItem -method in BlogItemController

+++ src/main/java/fi/tuni/blogproject/BlogItemController.java
@@ -17,6 +17,9 @@ public class BlogItemController {
     @Autowired
     CommentRepository commentRepository;
 
+    @Autowired
+    TagRepository tagRepository;
+
     /**
      * Adds a new BlogItem to database.
      *
@@ -35,8 +38,8 @@ public class BlogItemController {
      * Deletes a BlogItem from the database.
      *
      * <p>
-     *     Before the BlogItem is deleted, first deletes all the Comments
-     *     for this BlogItem if there are any.
+     *     Before the BlogItem is deleted, first deletes all the Comments and
+     *     tags, this this BlogItem has, if there are any.
      * </p>
      *
      * @param blogId Id of the BlogItem to be deleted.
@@ -45,13 +48,20 @@ public class BlogItemController {
     @DeleteMapping("/deleteBlogItem/{blogId}")
     @Transactional
     public Optional<BlogItem> deleteBlogItem(@PathVariable String blogId) {
-        for (Comment c : commentRepository.findAll()) {
-            if (c.getBlogId().equals(blogId)) {
-                commentRepository.deleteById(c.getId());
-            }
-        }
         Optional<BlogItem> bi = getBlogItem(blogId);
-        blogItemRepository.deleteById(blogId);
+        if (blogItemRepository.findById(blogId).isPresent()) {
+            for (Comment c : commentRepository.findAll()) {
+                if (c.getBlogId().equals(blogId)) {
+                    commentRepository.deleteById(c.getId());
+                }
+            }
+            for (Tag t : tagRepository.findAll()) {
+                if (t.getBlogId().equals(blogId)) {
+                    tagRepository.deleteById(t.getId());
+                }
+            }
+            blogItemRepository.deleteById(blogId);
+        }
         return bi;
     }
